requirements.txt
Django==4.0.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
asgiref>=3.5.2
pillow==2.2.1
boto3==1.22.9
botocore>=1.25.9
s3transfer==0.5.2